requirements.txt
Django==1.11.23
psycopg2-binary==2.8.4
python-dotenv==0.10.3
pytz==2017.2
